refactor: move server selection and max staleness tests to unit
